requirements.txt
streamlit>=1.32.0
pandas>=1.5.0
python-dotenv>=1.0.0
requests>=2.28.0
nltk>=3.8.0
nba_api>=1.4.0
